Allow non-integral average production times
Sample.averageProductionTimeMinutes is now a double instead of int, so
values like 2.5 minutes/unit are representable. Production duration is
rounded up (ceil) rather than truncated, since std::chrono::minutes has
no sub-minute precision and rounding down would let production appear
to finish early.

+++ SampleOrderSystem/Models/Sample.cpp
@@ -46,7 +46,7 @@ Sample Sample::FromJson(const JsonValue& json) {
     if (!averageProductionTimeMinutes.IsNumber()) {
         throw std::invalid_argument("Sample::FromJson: \"averageProductionTimeMinutes\" must be a number");
     }
-    sample.averageProductionTimeMinutes = static_cast<int>(averageProductionTimeMinutes.AsNumber());
+    sample.averageProductionTimeMinutes = averageProductionTimeMinutes.AsNumber();
 
     const JsonValue& yield = RequireField(json, "yield");
     if (!yield.IsNumber()) {
